requirements.txt
Flask==2.1.1
openai==0.27.0
python-dotenv==0.19.2
Flask-CORS==3.0.10
Werkzeug==2.0.3
